refactor: IAM 키 제거
