+++ src/main/java/com/example/kloset_lab/global/config/RestClientConfig.java
@@ -1,4 +1,4 @@
-package com.example.kloset_lab.global.security.config;
+package com.example.kloset_lab.global.config;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
