+++ src/main/resources/application-dev.yml
@@ -97,9 +97,6 @@ logging:
 
 cloud:
   aws:
-    credentials:
-      access-key: ${S3_ACCESS_KEY:none}
-      secret-key: ${S3_SECRET_KEY:none}
     region:
       static: ${S3_REGION:ap-northeast-2}
     s3:
