+++ src/main/resources/application-local.yml
@@ -15,12 +15,12 @@ jwt:
 cloud:
   aws:
     credentials:
-      access-key: ${S3_ACCESS_KEY}
-      secret-key: ${S3_SECRET_KEY}
+      access-key: ${AWS_ACCESS_KEY_ID}
+      secret-key: ${AWS_SECRET_ACCESS_KEY}
     region:
-      static: ${S3_REGION:ap-northeast-2}
+      static: ${AWS_REGION:ap-northeast-2}
     s3:
-      bucket: ${S3_BUCKET_NAME:}
+      bucket: ${AWS_S3_BUCKET:}
 
 kakao:
   client-id: ${KAKAO_CLIENT_ID}
